+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The public API of this library consists of the public functions declared in
 file [H3Core.java](./src/main/java/com/uber/h3core/H3Core.java), and support
 for the Linux x64 and Darwin x64 platforms.
 
+## Unreleased Changes
+### Fixed
+- Fixed the path to Windows resources.
+
 ## [4.0.0] - 2022-08-23
 ### Breaking Changes
 - Upgraded the core library to v4.0.0. (#104, #103, #102, #91)